+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## _Unreleased_
 
+- Added `-Dmaven.clean.failOnError=false` to the `mvn` template's default `clean` service
+so that running `lc clean` when scratch volumes are enabled won't cause the build to fail.
+
 - Deprecated `system upgrade`, since users should upgrade `lc` when they upgrade `lds`
 by running `lds upgrade`.